Organize project / absolute paths
