+++ components/button/AbsButton.tsx
@@ -1,10 +1,15 @@
+import { IAbsButton } from "App/types/global";
 import classNames from "classnames";
-import { IAbsButton } from "../../types/global";
 
-export const AbsButton = ({ handler = () => {}, optionalClass, children, radius = "rounded-full" }: IAbsButton) => {
+export default function AbsButton({
+  handler = () => {},
+  optionalClass,
+  children,
+  radius = "rounded-full",
+}: IAbsButton) {
   return (
     <button className={classNames("bg-el-3 dark:bg-el-3-dark", radius, optionalClass)} onClick={handler}>
       {children}
     </button>
   );
-};
+}
