+++ components/button/index.ts
@@ -1,0 +1,1 @@
+export { default as AbsButton } from "./AbsButton";
