+++ components/icons/svgs.tsx
@@ -1,5 +1,3 @@
-import classNames from "classnames";
-
 export const BiArrowSVG = (
   <svg
     xmlns="http://www.w3.org/2000/svg"
