+++ components/logo/index.ts
@@ -1,0 +1,1 @@
+export { default as AppLogo } from "./AppLogo";
