+++ components/mini-badge/index.ts
@@ -1,0 +1,1 @@
+export { default as MiniBadge } from "./MiniBadge";
