+++ components/navbar/NavLink.tsx
@@ -3,7 +3,6 @@
 import classNames from "classnames";
 import Link from "next/link";
 import { usePathname } from "next/navigation";
-import React from "react";
 
 export default function NavLink({ link, title }) {
   const pathname = usePathname();
