+++ components/table/index.ts
@@ -1,0 +1,2 @@
+export * from "./AbsTable";
+export { default as AbsTable } from "./AbsTable";
