+++ components/typography/SectionTitle.tsx
@@ -1,7 +1,7 @@
+import { ISectionTitle } from "App/types/global";
 import classNames from "classnames";
-import { ISectionTitle } from "../../types/global";
 
-export function SectionTitle({ children }: ISectionTitle) {
+export default function SectionTitle({ children }: ISectionTitle) {
   return (
     <p className={classNames("text-secondary dark:text-secondary-dark", "font-medium text-base", "mb-4")}>{children}</p>
   );
