+++ types/global.ts
