+++ utils/hash.ts
